Tile Searching Update
Changed how the game grid stored the tiles to help with searching tiles.

For now the game grid will be handling all the searching and getting of tiles in a board.

Tested with getting only the neighbouring tiles of the given tile on drop. This will be used to get the tiles to generate units for the building or 1 square distance movement.